Eliminar videos listo

+++ controlador/controladorVideos/fetch.js
@@ -4,8 +4,7 @@ export function peticionSubirVideo(datosVideo) {
     fetch('../../modelo/modeloVideos/subir-video.php', {
         method: 'POST',
         body: datosVideo,
-    })
-        .then((response) => response.json())
+    }).then((response) => response.json())
         .then((data) => {
             funciones.barraProgreso(data);
             const { mensaje } = data;
@@ -36,17 +35,16 @@ export function peticionSubirVideo(datosVideo) {
                 iconoAlerta = 'success';
             }
 
-            swal
-                .fire({
-                    title: tituloAlerta,
-                    text: mensajeAlerta,
-                    icon: iconoAlerta,
-                })
-                .then(() => {
-                    funciones.eliminarBarraProgreso();
-                    funciones.eliminarVideoRenderizado();
-                    peticionListarVideos();
-                });
+            swal.fire({
+                title: tituloAlerta,
+                text: mensajeAlerta,
+                icon: iconoAlerta,
+            }).then(() => {
+                funciones.eliminarBarraProgreso();
+                funciones.eliminarVideoRenderizado();
+                funciones.activarBotonPublicar();
+                peticionListarVideos();
+            });
         })
         .catch((err) => {
             console.log(err);
@@ -55,12 +53,44 @@ export function peticionSubirVideo(datosVideo) {
 
 export function peticionListarVideos() {
     fetch('../../modelo/modeloVideos/listar-videos.php')
-    .then(response => response.json())
-    .then((data) => {
-        if (data.mensaje === 'sin_videos') {
-            funciones.mensajeSinVideos('No hay vídeos por el momento');
-        } else {
-            funciones.renderizarListaVideos(data);
-        }
-    });
+        .then(response => response.json())
+        .then((data) => {
+            if (data.mensaje === 'sin_videos') {
+                funciones.eliminarListaVideos();
+                funciones.mensajeSinVideos('No hay vídeos por el momento');
+            } else {
+                funciones.renderizarListaVideos(data);
+            }
+        });
+}
+
+export function peticionEliminarVideo(datosVideo) {
+    fetch('../../modelo/modeloVideos/eliminar-video.php', {
+        method: 'POST',
+        body: datosVideo
+    }).then(response => response.json())
+        .then(data => {
+            let tituloAlerta,
+                mensajeAlerta,
+                iconoAlerta = '';
+
+            if (data.mensaje === 'video_eliminado') {
+                tituloAlerta = `Vídeo eliminado correctamente.`;
+                mensajeAlerta = `El vídeo "-${data.titulo}-" ha sido eliminado.`;
+                iconoAlerta = 'success';
+
+            } else if (data.mensaje === 'video_noeliminado') {
+                tituloAlerta = `Error al intentar eliminar`;
+                mensajeAlerta = `Ocurrió un error al intentar eliminar el vídeo "-${data.titulo}-"`;
+                iconoAlerta = 'error';
+            }
+
+            swal.fire({
+                title: tituloAlerta,
+                text: mensajeAlerta,
+                icon: iconoAlerta,
+            }).then(() => {
+                peticionListarVideos();
+            });
+        });
 }
